+++ client/router.js
@@ -6,8 +6,8 @@ Vue.use(Router)
 const Home = () => import('~/pages/home').then(m => m.default || m)
 const Login = () => import('~/pages/login').then(m => m.default || m)
 const Register = () => import('~/pages/register').then(m => m.default || m)
-// const PasswordReset = () => import('~/pages/auth/password/reset').then(m => m.default || m)
-// const PasswordRequest = () => import('~/pages/auth/password/email').then(m => m.default || m)
+const ForgotPassword = () => import('~/pages/forgot-password').then(m => m.default || m)
+const ResetPassword = () => import('~/pages/reset-password').then(m => m.default || m)
 
 // const Settings = () => import('~/pages/settings/index').then(m => m.default || m)
 // const SettingsProfile = () => import('~/pages/settings/profile').then(m => m.default || m)
@@ -16,9 +16,9 @@ const Register = () => import('~/pages/register').then(m => m.default || m)
 const routes = [
   { path: '/', name: 'home', component: Home },
   { path: '/login', name: 'login', component: Login },
-  { path: '/register', name: 'register', component: Register }
-  // { path: '/password/reset', name: 'password.request', component: PasswordRequest },
-  // { path: '/password/reset/:token', name: 'password.reset', component: PasswordReset },
+  { path: '/register', name: 'register', component: Register },
+  { path: '/forgot-password', name: 'forgot-password', component: ForgotPassword },
+  { path: '/reset-password/:token', name: 'password-reset', component: ResetPassword }
 
   // { path: '/settings', component: Settings, children: [
   //   { path: '', redirect: { name: 'settings.profile' }},